feat: 重构文件相关服务，解耦格式化和上传 feat: 新增保存一般文件并上传MinIO的功能 refactor: 修改数据库中消息id的存储类型，使其对应实际类型

+++ QQBotService/src/main/java/com/yachiyo/QQBotService/entity/GroupMessage.java
@@ -23,7 +23,7 @@ public class GroupMessage {
     private Long groupId; // 群号
 
     @TableField("message_id")
-    private Long messageId; // NapCat 消息ID
+    private Integer messageId; // NapCat 消息ID
 
     @TableField("sender_id")
     private Long senderId; // 发送者QQ号
